Turret Ability to fire and cooldown
Turrets can now call the fire function upon a created timer ticking and if tank is in range. Will need to give each turret its own timer that begins after firing and setting a CanFire bool to true after ticking, stopping the timer.
Also added some refactoring for checking if Tank is in range of turret.